Add the setup for commands
Commands must satisfy the command interface and will run as go routines
this gives the option for the command to operate in the background.
Currently there is not reason as the main thread will immediatly exit.
however controlling when the thread returns and not needing to return
yourself could allow for extra functionality down the road.

Main motivation for this is the ability to launch background processes
then return to main and continue on task in background.