Change Command Structure to use interfaces
Interfaces are now used to build out the commands. Executable makes
it possible to perform an action form the cli and Describable is
able to work with the help messages and printing syntax.

By convention a library should export a GetCommand function for the
use of the main cli to bring in the command and execute it by name.

Left todo is to standardize the naming and aliasing of commands by
providing some place to store information about the system and what
commands are avalible along with preferences.