Change to Stuct for Commands from Interface
The interface method has downfalls with enforcing the correct usage as
there can be no garentee that the implementer will follow conventions.
For this reason this attempts to convert the Command interface to a
struct. This will allow the same type to store the action as well as the
information about the command.

Main motivation for taking this approach is to reduce the need for type
conversion. Command struct will have all information for the main
program to operate off of.

Patch intoduces:
 - The help package is updated to run off the struct Command
 - The main program had an issue with no sub command error not printing
   if no args were passed as well as not exiting the program
 - Command package now only has the Command struct and instructions on
   adding new commands