GetCommand no longer needed in each Command
There is no need to have a set function for bringing in new commands
after looking into reflection capabilities of go there will not be a way
to invoke the method on a package not included at compiletime.

Instead the internal commands can just export a Command and main will
import and add it to a slice. This will allow the package to configure
the name of the sub command since main will access the name field on
the struct.